revert: findCommonPrefix optimisation
